audio/requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
ament-clang-format==0.12.11
ament-cmake-test==1.3.11
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
builtin-interfaces==1.2.1
catkin-pkg-modules==1.0.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.18.4
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.1
control-msgs==4.7.0
controller-manager==2.44.0
controller-manager-msgs==2.44.0
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==3.2.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.4
distlib==0.3.4
distro==1.9.0
distro-info==1.1+ubuntu0.2
docstring-parser==0.15
docutils==0.17.1
domain-coordinator==0.10.0
duplicity==0.8.21
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-guard-conditions==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
examples-rclpy-pointcloud-publisher==0.15.2
exceptiongroup==1.2.2
fasteners==0.14.1
flake8==4.0.1
fonttools==4.29.1
frozenlist==1.4.1
fs==2.4.12
future==0.18.2
gast==0.5.2
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
graph-msgs==0.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
iniconfig==1.1.1
instructor==0.6.8
interactive-markers==2.3.2
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.6.1
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
laser-geometry==2.4.0
launch==1.0.7
launch-param-builder==0.1.1
launch-ros==0.19.8
launch-testing==1.0.7
launch-testing-examples==0.15.2
launch-testing-ros==0.19.8
launch-xml==1.0.7
launch-yaml==1.0.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.1
lockfile==0.12.2
logging-demo==0.20.5
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
map-msgs==2.1.0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mdurl==0.1.2
message-filters==4.3.5
monotonic==1.6
more-itertools==8.10.0
moveit==0.1.3
moveit-configs-utils==2.5.5
moveit-msgs==2.2.1
moveit-task-constructor-msgs==0.1.3
mpi4py==3.1.3
mpmath==0.0.0
multidict==6.1.0
my-turtle-actions==0.0.0
nav-msgs==4.2.4
netifaces==0.11.0
nose2==0.9.2
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
olefile==0.46
openai==1.52.1
osrf-pycommon==2.1.4
packaging==21.3
paramiko==2.9.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
propcache==0.2.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
PyAudio==0.2.14
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.22
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pydocstyle==6.1.1
pydot==1.4.2
pydub==0.25.1
pyflakes==2.4.0
Pygments==2.18.0
PyGObject==3.42.1
pygraphviz==1.7
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-qt-binding==1.1.2
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.15
rcutils==5.1.6
reportlab==3.6.8
requests==2.25.1
resource-retriever==3.1.2
rich==13.9.3
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.11
ros2bag==0.15.13
ros2cli==0.18.11
ros2component==0.18.11
ros2controlcli==2.44.0
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.8
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2test==0.4.0
ros2topic==0.18.11
rosbag2-interfaces==0.15.13
rosbag2-py==0.15.13
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rospkg-modules==1.5.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
shell_gpt==1.4.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sounddevice==0.5.1
soupsieve==2.3.1
srdfdom==2.0.7
sros2==0.10.5
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.9
systemd-python==234
teleop-twist-keyboard==2.4.0
tenacity==8.5.0
tf2-geometry-msgs==0.25.9
tf2-kdl==0.25.9
tf2-msgs==0.25.9
tf2-py==0.25.9
tf2-ros-py==0.25.9
tf2-tools==0.25.9
toml==0.10.2
topic-monitor==0.20.5
tqdm==4.66.5
trajectory-msgs==4.2.4
turtlesim==1.4.2
typeguard==2.2.2
typer==0.9.4
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==1.26.5
usb-creator==0.3.7
vcstool==0.3.0
visualization-msgs==4.2.4
wadllib==1.3.6
webencodings==0.5.1
websockets==14.1
xacro==2.0.8
xdg==5
xkit==0.0.0
yarl==1.16.0
zipp==1.0.0
